Ensure just one and the right one post hook is executed.
If there are multiple transitions leading to single state, previous
logic could call incorrect post hook.
Now every transition sets the name of a posthook to call in the postSave
phase. The name of posthook is inserted in model variable
"state_machine_post_hook". In the postSave step, the variable is checked
by is_callable and called.